feat: don't require project ID
Documentation was misunderstood and this is not required as it is
already checked by Zitadel. In fact during local testing the two
applications were already residing in the same project which meant no
additional scopes had to be set. Application ID and Project ID were
confused by the author (me).